@@ -1,10 +1,10 @@
 {
     "require-dev": {
-        "symfony/phpunit-bridge": "6.0.0",
-        "vimeo/psalm": "4.18.1",
-        "psalm/plugin-phpunit": "0.16.1",
-        "weirdan/doctrine-psalm-plugin": "2.0.1",
-        "phpunit/phpunit": "9.5.11"
+        "symfony/phpunit-bridge": "6.1.3",
+        "vimeo/psalm": "4.25.0",
+        "psalm/plugin-phpunit": "0.17.0",
+        "weirdan/doctrine-psalm-plugin": "2.3.0",
+        "phpunit/phpunit": "9.5.21"
     },
     "config": {
         "allow-plugins": {
