@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Serendipity HQ Features Bundle.
  *
@@ -21,11 +23,8 @@ abstract class AbstractInvoiceDrawer implements InvoiceDrawerInterface
     /** @var string */
     public $locale;
 
-    /** @var \NumberFormatter */
-    private $currencyFormatter;
-
-    /** @var TranslatorInterface $translator */
-    private $translator;
+    private \NumberFormatter $currencyFormatter;
+    private TranslatorInterface $translator;
 
     public function __construct(TranslatorInterface $translator, string $locale)
     {
