@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Serendipity HQ Features Bundle.
  *
@@ -11,7 +13,6 @@
 
 namespace SerendipityHQ\Bundle\FeaturesBundle\Manager;
 
-use function Safe\sprintf;
 use SerendipityHQ\Bundle\FeaturesBundle\Form\DataTransformer\FeaturesCollectionTransformer;
 use SerendipityHQ\Bundle\FeaturesBundle\Form\Type\FeaturesType;
 use SerendipityHQ\Bundle\FeaturesBundle\Model\Feature\Configured\ConfiguredBooleanFeature;
@@ -25,7 +26,6 @@ use SerendipityHQ\Bundle\FeaturesBundle\Model\Feature\Property\IsRecurringFeatur
 use SerendipityHQ\Bundle\FeaturesBundle\Model\Feature\Subscribed\SubscribedBooleanFeature;
 use SerendipityHQ\Bundle\FeaturesBundle\Model\Feature\Subscribed\SubscribedCountableFeature;
 use SerendipityHQ\Bundle\FeaturesBundle\Model\Feature\Subscribed\SubscribedCountableFeaturePack;
-use SerendipityHQ\Bundle\FeaturesBundle\Model\Feature\Subscribed\SubscribedFeatureInterface;
 use SerendipityHQ\Bundle\FeaturesBundle\Model\Feature\Subscribed\SubscribedFeaturesCollection;
 use SerendipityHQ\Bundle\FeaturesBundle\Model\Feature\Subscribed\SubscribedRechargeableFeature;
 use SerendipityHQ\Bundle\FeaturesBundle\Model\Subscription;
@@ -35,6 +35,8 @@ use SerendipityHQ\Component\ValueObjects\Money\MoneyInterface;
 use Symfony\Component\Form\Extension\Core\Type\FormType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\Form\FormFactoryInterface;
+
+use function Safe\sprintf;
 
 /**
  * Contains method to manage features plans.
@@ -52,20 +54,12 @@ final class FeaturesManager
         SubscriptionInterface::MONTHLY  => 3,
         SubscriptionInterface::YEARLY   => 4,
     ];
-    /** @var ConfiguredFeaturesCollection $configuredFeatures */
-    private $configuredFeatures;
 
-    /** @var FormFactoryInterface $formFactory */
-    private $formFactory;
-
-    /** @var InvoicesManager $invoicesManager */
-    private $invoicesManager;
-
-    /** @var SubscriptionInterface $subscription */
-    private $subscription;
-
-    /** @var SubscriptionInterface $subscription This is use to calculate added and removed boolean features and the changed packs of CountableFeatures */
-    private $oldSubscription;
+    private ConfiguredFeaturesCollection $configuredFeatures;
+    private FormFactoryInterface $formFactory;
+    private InvoicesManager $invoicesManager;
+    private SubscriptionInterface $subscription;
+    private SubscriptionInterface $oldSubscription;
 
     /** @var array $differences The added and removed features */
     private $differences = [
@@ -93,9 +87,6 @@ final class FeaturesManager
         return $this->subscription;
     }
 
-    /**
-     * @return FeaturesManager
-     */
     public function setSubscription(SubscriptionInterface $subscription): self
     {
         $this->subscription = $subscription;
@@ -104,8 +95,6 @@ final class FeaturesManager
 
         /**
          * Set the Configured feature in each subscribed feature.
-         *
-         * @var SubscribedFeatureInterface
          */
         foreach ($subscription->getFeatures()->getValues() as $subscribedFeature) {
             $configuredFeature = $this->getConfiguredFeatures()->get($subscribedFeature->getName());
@@ -127,18 +116,11 @@ final class FeaturesManager
         return $this;
     }
 
-    /**
-     * @throws \InvalidArgumentException If the $subscriptionInterval does not exist
-     */
     public function buildDefaultSubscriptionFeatures(string $subscriptionInterval): SubscribedFeaturesCollection
     {
         $activeUntil = Subscription::calculateActiveUntil($subscriptionInterval);
         $features    = [];
 
-        /**
-         * @var string
-         * @var ConfiguredBooleanFeature|ConfiguredCountableFeature|ConfiguredRechargeableFeature|FeatureInterface $details
-         */
         foreach ($this->getConfiguredFeatures() as $name => $details) {
             switch ($details->getType()) {
                 case FeatureInterface::TYPE_BOOLEAN:
@@ -178,10 +160,8 @@ final class FeaturesManager
     /**
      * @param SubscribedFeaturesCollection $newFeatures This comes from the form, not from the Subscription! The Subscription is
      *                                                  not yet synced with these new Features!
-     *
-     * @return Money
      */
-    public function calculateTotalChargesForNewFeatures(SubscribedFeaturesCollection $newFeatures)
+    public function calculateTotalChargesForNewFeatures(SubscribedFeaturesCollection $newFeatures): MoneyInterface
     {
         $totalCharges = new Money([MoneyInterface::BASE_AMOUNT => 0, MoneyInterface::CURRENCY => $this->getSubscription()->getCurrency()]);
 
@@ -213,6 +193,7 @@ final class FeaturesManager
                             // If it continues to work as expected, remove this entire comment.
                             break;
                         }
+
                         $price = $configuredFeature->getInstantPrice($this->getSubscription()->getCurrency(), $this->getSubscription()->getRenewInterval(), FeatureInterface::PRICE_GROSS);
 
                         break;
@@ -228,12 +209,10 @@ final class FeaturesManager
                         }
 
                         break;
-                    // A RechargeableFeature hasn't a subscription period, so it hasn't an isStillActive() method
+                        // A RechargeableFeature hasn't a subscription period, so it hasn't an isStillActive() method
                     case SubscribedRechargeableFeature::class:
                         /**
                          * For the moment force the code to get the pack's instant price.
-                         *
-                         * @var SubscribedRechargeableFeature
                          */
                         $price = $configuredFeature->getPack($checkingFeature->getRechargingPack()->getNumOfUnits())->getPrice($this->getSubscription()->getCurrency(), FeatureInterface::PRICE_GROSS);
 
@@ -344,8 +323,6 @@ final class FeaturesManager
         if (null !== $newFeatures) {
             /**
              * Before all, update the features, setting the new enabled status or adding the feature if not already present.
-             *
-             * @var FeatureInterface
              */
             foreach ($newFeatures as $newFeature) {
                 $existentFeature = $this->getSubscription()->getFeatures()->get($newFeature->getName());
@@ -455,8 +432,6 @@ final class FeaturesManager
          * A feature is removed if:
          * 1. It was in the old collection but doesn't exist in the new collection;
          * 2. It was in the old collection and was enabled and is in the new collection but is not enabled
-         *
-         * @var FeatureInterface
          */
         foreach ($oldFeatures as $oldFeature) {
             // If the Feature is in the old collection but doesn't exist in the new collection...
@@ -481,19 +456,14 @@ final class FeaturesManager
                     }
 
                     break;
-                // If is a CountableFeature...
+                    // If is a CountableFeature...
                 case SubscribedCountableFeature::class:
                     /**
                      * ... and was in the old collection and in the new collection, too ...
-                     *
-                     * @var SubscribedCountableFeature
                      */
                     if (true === $newFeatures->containsKey($oldFeature->getName())) {
                         /**
                          * We first get the subscribed packages...
-                         *
-                         * @var SubscribedCountableFeaturePack
-                         * @var SubscribedCountableFeaturePack $newSubscribedPack
                          */
                         $oldSubscribedPack = $oldFeature->getSubscribedPack();
                         $newSubscribedPack = $newFeatures->get($oldFeature->getName())->getSubscribedPack();
@@ -517,8 +487,6 @@ final class FeaturesManager
          * A feature is added if:
          * 1. It was not in the old collection but exists in the new collection;
          * 2. It was in the old collection and was not enabled and is in the new collection too but is enabled
-         *
-         * @var SubscribedBooleanFeature|SubscribedCountableFeature|SubscribedRechargeableFeature
          */
         foreach ($newFeatures as $newFeature) {
             /*
@@ -534,13 +502,13 @@ final class FeaturesManager
                     $featureDetails = $newFeature->getName();
 
                     break;
-                // If is a CountableFeature...
+                    // If is a CountableFeature...
                 case SubscribedCountableFeature::class:
                     /** @var SubscribedCountableFeature $featureDetails */
                     $featureDetails = [$newFeature->getName() => $newFeature->getSubscribedPack()->getNumOfUnits()];
 
                     break;
-                // If is a CountableFeature...
+                    // If is a CountableFeature...
                 case SubscribedRechargeableFeature::class:
                     /** @var SubscribedRechargeableFeature $featureDetails */
                     $featureDetails = [$newFeature->getName() => $newFeature->getRechargingPack()->getNumOfUnits()];
@@ -572,7 +540,7 @@ final class FeaturesManager
                         }
 
                         break;
-                    // If is a CountableFeature...
+                        // If is a CountableFeature...
                     case SubscribedCountableFeature::class:
                         /** @var SubscribedCountableFeaturePack $newSubscribedPack */
                         $newSubscribedPack = $newFeature->getSubscribedPack();
@@ -587,7 +555,7 @@ final class FeaturesManager
                         }
 
                         break;
-                    // If it is a RechargeableFeature...
+                        // If it is a RechargeableFeature...
                     case SubscribedRechargeableFeature::class:
                         // ... if a rechargin pack exists...
                         if ($newFeature->hasRechargingPack()) {
@@ -634,6 +602,7 @@ final class FeaturesManager
                 $feature->refresh();
             }
         }
+
         $nextRefreshOn = $this->getSubscription()->getNextRefreshOn() ?? clone $this->getSubscription()->getSubscribedOn();
         $this->getSubscription()
             ->setSmallestRefreshInterval($refreshInterval)
