@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Serendipity HQ Features Bundle.
  *
@@ -13,11 +15,8 @@ namespace SerendipityHQ\Bundle\FeaturesBundle\Model\Feature;
 
 abstract class AbstractFeature implements FeatureInterface
 {
-    /** @var string $name */
-    private $name;
-
-    /** @var string $type */
-    private $type;
+    private string $name;
+    private string $type;
 
     public function __construct(string $name, array $details = [])
     {
