@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Serendipity HQ Features Bundle.
  *
@@ -32,11 +34,8 @@ final class ConfiguredCountableFeature extends AbstractFeature implements HasCon
     /** @var string $refreshPeriod */
     private $refreshPeriod;
 
-    /** @var string $taxName */
-    private $taxName;
-
-    /** @var float $taxRate */
-    private $taxRate;
+    private string $taxName;
+    private float $taxRate;
 
     public function __construct(string $name, array $details = [])
     {
