@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Serendipity HQ Features Bundle.
  *
@@ -11,34 +13,28 @@
 
 namespace SerendipityHQ\Bundle\FeaturesBundle\Model\Feature\Subscribed;
 
-use function Safe\json_decode;
-use function Safe\json_encode;
-use function Safe\sprintf;
 use SerendipityHQ\Bundle\FeaturesBundle\Model\Feature\Property\CanBeConsumedInterface;
 use SerendipityHQ\Bundle\FeaturesBundle\Model\Feature\Property\CanBeConsumedProperty;
 
-/**
- * {@inheritdoc}
- */
+use function Safe\json_decode;
+use function Safe\json_encode;
+use function Safe\sprintf;
+
 final class SubscribedRechargeableFeature extends AbstractSubscribedFeature implements SubscribedFeatureInterface, CanBeConsumedInterface
 {
     use CanBeConsumedProperty;
 
-    /** @var \DateTime $lastRecharge The last time a recharge was done */
-    private $lastRechargeOn;
+    /** @var \DateTimeInterface $lastRecharge The last time a recharge was done */
+    private \DateTimeInterface $lastRechargeOn;
 
     /** @var int $lastRechargeQuantity The quantity of units recharged last time */
-    private $lastRechargeQuantity;
+    private int $lastRechargeQuantity;
 
-    /** @var SubscribedRechargeableFeaturePack */
-    private $rechargingPack;
+    private SubscribedRechargeableFeaturePack $rechargingPack;
 
     /** @var int $remainedQuantity The amount of remained units */
     private $remainedQuantity = 0;
 
-    /**
-     * {@inheritdoc}
-     */
     public function __construct(string $name, array $details = [])
     {
         // Set the type
@@ -56,8 +52,6 @@ final class SubscribedRechargeableFeature extends AbstractSubscribedFeature impl
     }
 
     /**
-     * {@inheritdoc}
-     *
      * @return \DateTime|\DateTimeImmutable
      */
     public function getLastRechargeOn(): \DateTimeInterface
@@ -65,17 +59,11 @@ final class SubscribedRechargeableFeature extends AbstractSubscribedFeature impl
         return $this->lastRechargeOn;
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function getLastRechargeQuantity(): int
     {
         return $this->lastRechargeQuantity;
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function getRechargingPack(): SubscribedRechargeableFeaturePack
     {
         if (false === $this->hasRechargingPack()) {
@@ -85,17 +73,11 @@ final class SubscribedRechargeableFeature extends AbstractSubscribedFeature impl
         return $this->rechargingPack;
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function hasRechargingPack(): bool
     {
         return null !== $this->rechargingPack;
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function recharge(): SubscribedRechargeableFeature
     {
         $rechargeQuantity = $this->getRechargingPack()->getNumOfUnits();
@@ -106,9 +88,6 @@ final class SubscribedRechargeableFeature extends AbstractSubscribedFeature impl
         return $this;
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function setRecharginPack(SubscribedRechargeableFeaturePack $rechargingPack): SubscribedRechargeableFeature
     {
         $this->rechargingPack = $rechargingPack;
@@ -116,9 +95,6 @@ final class SubscribedRechargeableFeature extends AbstractSubscribedFeature impl
         return $this;
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function toArray(): array
     {
         return \array_merge([
