@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Serendipity HQ Features Bundle.
  *
@@ -13,9 +15,10 @@ namespace SerendipityHQ\Bundle\FeaturesBundle\Model;
 
 use Doctrine\ORM\Mapping as ORM;
 use Money\Currency;
-use function Safe\sprintf;
 use SerendipityHQ\Component\ValueObjects\Money\Money;
 use SerendipityHQ\Component\ValueObjects\Money\MoneyInterface;
+
+use function Safe\sprintf;
 
 /**
  * @ORM\MappedSuperclass
@@ -24,41 +27,25 @@ use SerendipityHQ\Component\ValueObjects\Money\MoneyInterface;
 abstract class Invoice implements InvoiceInterface
 {
     private const SECTION_DEFAULT = '_default';
-    /**
-     * @var Currency
-     * @ORM\Column(name="currency", type="currency")
-     */
-    private $currency;
+
+    /** @ORM\Column(name="currency", type="currency") */
+    private Currency $currency;
+
+    /** @ORM\Column(name="issued_on", type="datetime") */
+    private \DateTimeInterface $issuedOn;
 
     /**
-     * @var \DateTime
-     * @ORM\Column(name="issued_on", type="datetime")
-     */
-    private $issuedOn;
-
-    /**
-     * @var array
+     * @var InvoiceSection[]
      * @ORM\Column(name="`sections`", type="json")
      */
-    private $sections;
+    private array $sections = [];
 
-    /**
-     * @var MoneyInterface
-     *
-     * @ORM\Column(name="gross_total", type="money")
-     */
-    private $grossTotal;
+    /** @ORM\Column(name="gross_total", type="money") */
+    private MoneyInterface $grossTotal;
 
-    /**
-     * @var MoneyInterface
-     *
-     * @ORM\Column(name="net_total", type="money")
-     */
-    private $netTotal;
+    /** @ORM\Column(name="net_total", type="money") */
+    private MoneyInterface $netTotal;
 
-    /**
-     * {@inheritdoc}
-     */
     public function __construct($currency)
     {
         if ( ! $currency instanceof Currency) {
@@ -66,7 +53,6 @@ abstract class Invoice implements InvoiceInterface
         }
 
         $this->currency = $currency;
-        $this->sections = [];
 
         // Set the issue date
         if (null === $this->issuedOn) {
@@ -86,41 +72,26 @@ abstract class Invoice implements InvoiceInterface
         $this->generateNumber();
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function getHeader(): ?InvoiceSectionHeader
     {
         return $this->sections[self::SECTION_DEFAULT]->getHeader();
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function hasHeader(): bool
     {
         return $this->sections[self::SECTION_DEFAULT]->hasHeader();
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function removeHeader()
+    public function removeHeader(): InvoiceSection
     {
         return $this->sections[self::SECTION_DEFAULT]->removeHeader();
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function setHeader(InvoiceSectionHeader $header)
     {
         $this->sections[self::SECTION_DEFAULT]->setHeader($header);
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function addLine(InvoiceLine $line, string $id = null): InvoiceInterface
     {
         if (false === isset($this->sections[self::SECTION_DEFAULT])) {
@@ -134,33 +105,21 @@ abstract class Invoice implements InvoiceInterface
         return $this;
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function getLine($id): InvoiceLine
     {
         return $this->sections[self::SECTION_DEFAULT]->getLine($id);
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function getLines(): array
     {
         return $this->sections[self::SECTION_DEFAULT]->getLines();
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function hasLine($id): bool
     {
         return $this->sections[self::SECTION_DEFAULT]->hasLine($id);
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function removeLine($id)
     {
         $return = $this->sections[self::SECTION_DEFAULT]->removeLine($id);
@@ -170,9 +129,6 @@ abstract class Invoice implements InvoiceInterface
         return $return;
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function addSection(InvoiceSection $section, string $id = null): InvoiceInterface
     {
         if ($this->getCurrency()->getCode() !== $section->getCurrency()->getCode()) {
@@ -204,9 +160,6 @@ abstract class Invoice implements InvoiceInterface
         return $this;
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function getSection($id): ?InvoiceSection
     {
         if (self::SECTION_DEFAULT === $id && false === isset($this->sections[self::SECTION_DEFAULT])) {
@@ -216,17 +169,11 @@ abstract class Invoice implements InvoiceInterface
         return $this->sections[$id] ?? null;
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function getSections(): array
     {
         return $this->sections;
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function hasSection($id): bool
     {
         if (false === \is_string($id) && false === \is_int($id)) {
@@ -236,9 +183,6 @@ abstract class Invoice implements InvoiceInterface
         return isset($this->sections[$id]);
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function removeSection($id)
     {
         if ($this->hasSection($id)) {
@@ -253,17 +197,12 @@ abstract class Invoice implements InvoiceInterface
         return false;
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function getCurrency(): Currency
     {
         return $this->currency;
     }
 
     /**
-     * {@inheritdoc}
-     *
      * @return \DateTime|\DateTimeImmutable
      */
     public function getIssuedOn(): \DateTimeInterface
@@ -271,33 +210,22 @@ abstract class Invoice implements InvoiceInterface
         return $this->issuedOn;
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function getGrossTotal(): MoneyInterface
     {
         return $this->grossTotal;
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function getNetTotal(): MoneyInterface
     {
         return $this->netTotal;
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function jsonSerialize(): array
     {
         return $this->sections;
     }
 
     /**
-     * {@inheritdoc}
-     *
      * @ORM\PostLoad()
      */
     public function jsonUnserialize()
