@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Serendipity HQ Features Bundle.
  *
@@ -12,29 +14,24 @@
 namespace SerendipityHQ\Bundle\FeaturesBundle\Model;
 
 use Money\Currency;
-use function Safe\sprintf;
 use SerendipityHQ\Component\ValueObjects\Money\Money;
 use SerendipityHQ\Component\ValueObjects\Money\MoneyInterface;
+
+use function Safe\sprintf;
 
 /**
  * Represents a section into the Invoice.
  */
 final class InvoiceSection implements \JsonSerializable
 {
-    /** @var Currency $currency */
-    private $currency;
-
-    /** @var InvoiceSectionHeader */
-    private $header;
+    private Currency $currency;
+    private InvoiceSectionHeader $header;
 
     /** @var InvoiceLine[] */
-    private $lines = [];
+    private array $lines = [];
 
-    /** @var MoneyInterface $grossTotal */
-    private $grossTotal;
-
-    /** @var MoneyInterface $netTotal */
-    private $netTotal;
+    private MoneyInterface $grossTotal;
+    private MoneyInterface $netTotal;
 
     public function __construct(Currency $currency)
     {
@@ -153,9 +150,6 @@ final class InvoiceSection implements \JsonSerializable
         return $this->netTotal;
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function jsonSerialize(): array
     {
         $return = [];
