@@ -1,5 +1,16 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file is part of the Serendipity HQ Features Bundle.
+ *
+ * Copyright (c) Adamo Aerendir Crespi <aerendir@serendipityhq.com>.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 /**
  * This configuration will be read and overlaid on top of the
  * default configuration. Command line arguments will be applied
@@ -7,7 +18,7 @@
  */
 return [
     'target_php_version' => '7.4',
-    'minimum_severity' => \Phan\Issue::SEVERITY_LOW,
+    'minimum_severity'   => \Phan\Issue::SEVERITY_LOW,
 
     // A list of directories that should be parsed for class and
     // method information. After excluding the directories
@@ -20,7 +31,7 @@ return [
         'src',
         'tests',
         'vendor',
-        'vendor-bin/phpunit/vendor'
+        'vendor-bin/phpunit/vendor',
     ],
 
     // A directory list that defines files that will be excluded
@@ -34,22 +45,22 @@ return [
     //       party code, directories containing that code
     //       should be added to both the `directory_list`
     //       and `exclude_analysis_directory_list` arrays.
-    "exclude_analysis_directory_list" => [
+    'exclude_analysis_directory_list' => [
         'vendor/',
         'vendor-bin/phpunit/vendor',
         'build/',
-        'docs/'
+        'docs/',
     ],
 
-    'quick_mode' => false,
-    'analyze_signature_compatibility' => true,
-    'allow_missing_properties' => false,
-    'null_casts_as_any_type' => false,
-    'null_casts_as_array' => false,
-    'array_casts_as_null' => false,
-    'scalar_implicit_cast' => false,
+    'quick_mode'                                  => false,
+    'analyze_signature_compatibility'             => true,
+    'allow_missing_properties'                    => false,
+    'null_casts_as_any_type'                      => false,
+    'null_casts_as_array'                         => false,
+    'array_casts_as_null'                         => false,
+    'scalar_implicit_cast'                        => false,
     'ignore_undeclared_variables_in_global_scope' => false,
-    'suppress_issue_types' => [
+    'suppress_issue_types'                        => [
         'PhanUnreferencedPublicMethod',
         'PhanUnreferencedUseNormal',
     ],
@@ -62,9 +73,9 @@ return [
     // can't be removed for whatever reason.
     // (e.g. '@Test\.php$@', or '@vendor/.*/(tests|Tests)/@')
     'exclude_file_regex' => '@^vendor/.*/(tests?|Tests?)/@',
-    'plugins' => [
+    'plugins'            => [
         'vendor-bin/phan/vendor/drenso/phan-extensions/Plugin/Annotation/SymfonyAnnotationPlugin.php',
         'vendor-bin/phan/vendor/drenso/phan-extensions/Plugin/DocComment/InlineVarPlugin.php',
-        'vendor-bin/phan/vendor/drenso/phan-extensions/Plugin/DocComment/MethodPlugin.php'
-    ]
+        'vendor-bin/phan/vendor/drenso/phan-extensions/Plugin/DocComment/MethodPlugin.php',
+    ],
 ];
