@@ -1,9 +1,9 @@
 <?php
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 /*
- * This file is part of the Serendipity HQ Aws Ses Bundle.
+ * This file is part of the Serendipity HQ Features Bundle.
  *
  * Copyright (c) Adamo Aerendir Crespi <aerendir@serendipityhq.com>.
  *
@@ -11,69 +11,19 @@ declare(strict_types = 1);
  * file that was distributed with this source code.
  */
 
-use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
-use Rector\Core\Configuration\Option;
+use Rector\Config\RectorConfig;
 use Rector\Core\ValueObject\PhpVersion;
-use Rector\Set\ValueObject\SetList;
+use SerendipityHQ\Integration\Rector\SerendipityHQ;
 
-return static function (ContainerConfigurator $containerConfigurator) : void {
-    $parameters = $containerConfigurator->parameters();
-
-    $parameters->set(Option::PHP_VERSION_FEATURES, PhpVersion::PHP_74);
-
-    $parameters->set(Option::PATHS, [
-        __DIR__ . '/src',
-        __DIR__ . '/tests'
-    ]);
+return static function (RectorConfig $rectorConfig): void {
+    $rectorConfig->phpVersion(PhpVersion::PHP_74);
+    $rectorConfig->paths([__DIR__ . '/src', __DIR__ . '/tests']);
 
     // This causes issues with controllers
     // Until required for tests, keep it commented
-    // $parameters->set(Option::BOOTSTRAP_FILES, [__DIR__ . '/vendor-bin/phpunit/vendor/autoload.php']);
+    // $rectorConfig->bootstrapFiles([__DIR__ . '/vendor-bin/phpunit/vendor/autoload.php']);
+    $rectorConfig->import(SerendipityHQ::SHQ_SYMFONY_BUNDLE);
 
-    $containerConfigurator->import(SetList::ACTION_INJECTION_TO_CONSTRUCTOR_INJECTION);
-    $containerConfigurator->import(SetList::CODE_QUALITY);
-    $containerConfigurator->import(SetList::CODING_STYLE);
-    $containerConfigurator->import(SetList::MONOLOG_20);
-    $containerConfigurator->import(SetList::FRAMEWORK_EXTRA_BUNDLE_40);
-    $containerConfigurator->import(SetList::FRAMEWORK_EXTRA_BUNDLE_50);
-    $containerConfigurator->import(SetList::PHP_52);
-    $containerConfigurator->import(SetList::PHP_53);
-    $containerConfigurator->import(SetList::PHP_54);
-    $containerConfigurator->import(SetList::PHP_56);
-    $containerConfigurator->import(SetList::PHP_70);
-    $containerConfigurator->import(SetList::PHP_71);
-    $containerConfigurator->import(SetList::PHP_72);
-    $containerConfigurator->import(SetList::PHP_73);
-    $containerConfigurator->import(SetList::UNWRAP_COMPAT);
-    $containerConfigurator->import(SetList::SAFE_07);
-    $containerConfigurator->import(SetList::TYPE_DECLARATION);
-
-    $parameters->set(Option::AUTO_IMPORT_NAMES, true);
-    $parameters->set(Option::IMPORT_DOC_BLOCKS, true);
-    $parameters->set(Option::IMPORT_SHORT_CLASSES, false);
-
-    $parameters->set(
-        Option::SKIP,
-        [
-            Rector\CodeQuality\Rector\Catch_\ThrowWithPreviousExceptionRector::class,
-            Rector\CodeQuality\Rector\Concat\JoinStringConcatRector::class,
-            Rector\CodeQuality\Rector\Identical\SimplifyBoolIdenticalTrueRector::class,
-            Rector\CodingStyle\Rector\Class_\AddArrayDefaultToArrayPropertyRector::class,
-            Rector\CodingStyle\Rector\ClassConst\VarConstantCommentRector::class,
-            Rector\CodingStyle\Rector\ClassMethod\NewlineBeforeNewAssignSetRector::class,
-            Rector\CodingStyle\Rector\ClassMethod\RemoveDoubleUnderscoreInMethodNameRector::class,
-            Rector\CodingStyle\Rector\Encapsed\EncapsedStringsToSprintfRector::class,
-            Rector\CodingStyle\Rector\Switch_\BinarySwitchToIfElseRector::class,
-            Rector\Php56\Rector\FunctionLike\AddDefaultValueForUndefinedVariableRector::class, // Maybe good one day
-            Rector\PHPUnit\Rector\Class_\AddSeeTestAnnotationRector::class,
-            Rector\PHPUnit\Rector\ClassMethod\AddDoesNotPerformAssertionToNonAssertingTestRector::class,
-            Rector\Privatization\Rector\MethodCall\PrivatizeLocalGetterToPropertyRector::class,
-            Rector\TypeDeclaration\Rector\ClassMethod\AddArrayParamDocTypeRector::class,
-            Rector\TypeDeclaration\Rector\ClassMethod\AddArrayReturnDocTypeRector::class,
-            \Rector\CodingStyle\Rector\Stmt\NewlineAfterStatementRector::class,
-
-            // Bugged or to understand deeper
-            Rector\TypeDeclaration\Rector\FunctionLike\ReturnTypeDeclarationRector::class,
-        ]
-    );
+    $toSkip = SerendipityHQ::buildToSkip(SerendipityHQ::SHQ_SYMFONY_BUNDLE_SKIP);
+    $rectorConfig->skip($toSkip);
 };
